Feat: Add ability to generate HTML report
Add the ability to generate a HTML based report.
Change the flags so it's obvious which report is being generated.
Removed diff from JSON report and replace with the response bodies for candidate and control.
Add additional match property to JSON report.
Add additional status.metadata.eitherPathErrored property to JSON report.
Alter prettier settings so it does not break lines so often.